requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.6.3
pydantic>=2.6.4
motor==3.3.1
firebase-admin==6.2.0
google-cloud-firestore>=2.19.0
starlette>=0.36.0
httpx>=0.24.0
python-multipart>=0.0.9
